+++ msmexplorer/plots/histogram.py
@@ -1,3 +1,0 @@
-from corner import corner as plot_histogram
-
-__all__ = ['plot_histogram']