+++ docs/changelog.rst
@@ -10,5 +10,7 @@ This is the current development version of MSMExplorer
 
 New Features
 ~~~~~~~~~~~~
-+ Featuring plotting tools for clustering, chord diagrams,
-  Markov state modeling, and transition path theory.
++ Clustering plots: ``plot_voronoi``
++ MSM plots: ``plot_pop_resids``, ``plot_msm_network``, ``plot_timescales``
++ Projection plots: ``plot_free_energy``, ``plot_histogram``
++ TPT plots: ``plot_tpaths``